+++ p2p/src/main/kotlin/maru/p2p/MaruPeerHandler.kt
@@ -1,0 +1,29 @@
+/*
+   Copyright 2025 Consensys Software Inc.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+ */
+package maru.p2p
+
+import tech.pegasys.teku.networking.p2p.network.PeerHandler
+import tech.pegasys.teku.networking.p2p.peer.Peer
+
+class MaruPeerHandler : PeerHandler {
+  override fun onConnect(peer: Peer?) {
+    // TODO("Not yet implemented1")
+  }
+
+  override fun onDisconnect(peer: Peer?) {
+    // TODO("Not yet implemented2")
+  }
+}
