+++ p2p/src/main/kotlin/maru/p2p/P2PNetworkBuilder.kt
@@ -24,7 +24,6 @@ import io.libp2p.core.dsl.BuilderJ
 import io.libp2p.core.dsl.hostJ
 import io.libp2p.core.multiformats.Multiaddr
 import io.libp2p.core.mux.StreamMuxer
-import io.libp2p.core.pubsub.ValidationResult
 import io.libp2p.etc.types.millis
 import io.libp2p.protocol.PingProtocol
 import io.libp2p.pubsub.PubsubApiImpl
@@ -37,8 +36,6 @@ import io.libp2p.pubsub.gossip.builders.GossipRouterBuilder
 import io.libp2p.security.secio.SecIoSecureChannel
 import io.libp2p.transport.ConnectionUpgrader
 import io.libp2p.transport.tcp.TcpTransport
-import io.netty.buffer.ByteBuf
-import io.netty.buffer.ByteBufUtil
 import java.util.Optional
 import kotlin.random.Random
 import org.apache.tuweni.bytes.Bytes
@@ -46,7 +43,6 @@ import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem
 import pubsub.pb.Rpc
 import tech.pegasys.teku.infrastructure.async.AsyncRunnerFactory
 import tech.pegasys.teku.infrastructure.async.MetricTrackingExecutorFactory
-import tech.pegasys.teku.infrastructure.async.SafeFuture
 import tech.pegasys.teku.infrastructure.unsigned.UInt64
 import tech.pegasys.teku.networking.p2p.gossip.PreparedGossipMessage
 import tech.pegasys.teku.networking.p2p.gossip.TopicHandler
@@ -59,13 +55,8 @@ import tech.pegasys.teku.networking.p2p.libp2p.gossip.PreparedPubsubMessage
 import tech.pegasys.teku.networking.p2p.libp2p.rpc.RpcHandler
 import tech.pegasys.teku.networking.p2p.network.P2PNetwork
 import tech.pegasys.teku.networking.p2p.network.PeerHandler
-import tech.pegasys.teku.networking.p2p.peer.NodeId
 import tech.pegasys.teku.networking.p2p.peer.Peer
 import tech.pegasys.teku.networking.p2p.reputation.ReputationManager
-import tech.pegasys.teku.networking.p2p.rpc.RpcMethod
-import tech.pegasys.teku.networking.p2p.rpc.RpcRequestHandler
-import tech.pegasys.teku.networking.p2p.rpc.RpcResponseHandler
-import tech.pegasys.teku.networking.p2p.rpc.RpcStream
 
 private const val ORIGINAL_MESSAGE = "68656c6c6f"
 
@@ -74,6 +65,10 @@ open class P2PNetworkBuilder {
   private var ipv4Address: Multiaddr? = null
   private var ipv6Address: Multiaddr? = null
   private var peerManager: PeerManager? = null
+
+  companion object {
+    var rpcMethod = MaruRpcMethod()
+  }
 
   fun build(): P2PNetwork<Peer> {
     // check everything that is needed is set
@@ -249,182 +244,5 @@ open class P2PNetworkBuilder {
         b.protocols.add(rpcHandler)
       }
     return host
-  }
-
-  class MaruPeerHandler : PeerHandler {
-    override fun onConnect(peer: Peer?) {
-      // TODO("Not yet implemented1")
-    }
-
-    override fun onDisconnect(peer: Peer?) {
-      // TODO("Not yet implemented2")
-    }
-  }
-
-  open class MaruRpcMethod : RpcMethod<MaruOutgoingRpcRequestHandler, Bytes, MaruRpcResponseHandler> {
-    private lateinit var peerManager: PeerManager
-
-    fun setPeerManager(peerManager: PeerManager) {
-      this.peerManager = peerManager
-    }
-
-    override fun getIds(): MutableList<String> = mutableListOf("linea")
-
-    override fun createIncomingRequestHandler(p0: String?): RpcRequestHandler {
-      val maruRpcRequestHandler = MaruIncomingRpcRequestHandler(peerManager)
-      println("createIncomingRequestHandler $p0 $maruRpcRequestHandler")
-      return maruRpcRequestHandler
-    }
-
-    override fun createOutgoingRequestHandler(
-      protocolId: String?,
-      request: Bytes?,
-      responseHandler: MaruRpcResponseHandler?,
-    ): MaruOutgoingRpcRequestHandler {
-      val maruRpcRequestHandler = MaruOutgoingRpcRequestHandler(responseHandler!!, peerManager)
-      println("createOutgoingRequestHandler $protocolId $request $maruRpcRequestHandler")
-      return maruRpcRequestHandler
-    }
-
-    override fun encodeRequest(p0: Bytes?): Bytes = p0!!
-  }
-
-  open class MaruIncomingRpcRequestHandler(
-    private val peerManager: PeerManager,
-  ) : RpcRequestHandler {
-    private var data: String = ""
-
-    override fun active(
-      p0: NodeId?,
-      p1: RpcStream?,
-    ) {
-      println("active $this")
-    }
-
-    override fun processData(
-      p0: NodeId?,
-      p1: RpcStream?,
-      p2: ByteBuf?,
-    ) {
-      val bytes = ByteBufUtil.getBytes(p2!!)
-      println("processData request: $bytes ($this)")
-      data += bytes
-      p1!!.writeBytes(Bytes.wrap(bytes).reverse())
-    }
-
-    override fun readComplete(
-      p0: NodeId?,
-      p1: RpcStream?,
-    ) {
-      println("readComplete $data")
-    }
-
-    override fun closed(
-      p0: NodeId?,
-      p1: RpcStream?,
-    ) {
-      println("closed $data")
-    }
-  }
-
-  open class MaruOutgoingRpcRequestHandler(
-    private val responseHandler: MaruRpcResponseHandler,
-    private val peerManager: PeerManager,
-  ) : RpcRequestHandler {
-    private var data: String = ""
-
-    override fun active(
-      p0: NodeId?,
-      p1: RpcStream?,
-    ) {
-      println("active $this")
-    }
-
-    override fun processData(
-      p0: NodeId?,
-      p1: RpcStream?,
-      p2: ByteBuf?,
-    ) {
-      val bytes = ByteBufUtil.getBytes(p2!!)
-      val peer = peerManager.getPeer(p0).get()
-      println("processData response: $bytes from peer $peer.id ($this)")
-      data += bytes
-      p1!!.closeWriteStream()
-      responseHandler.onResponse(Bytes.wrap(bytes))
-    }
-
-    override fun readComplete(
-      p0: NodeId?,
-      p1: RpcStream?,
-    ) {
-      println("readComplete $data")
-    }
-
-    override fun closed(
-      p0: NodeId?,
-      p1: RpcStream?,
-    ) {
-      println("closed $data")
-    }
-  }
-
-  class TestTopicHandler : TopicHandler {
-    companion object {
-      val dataFuture = SafeFuture<Bytes>()
-    }
-
-    override fun prepareMessage(
-      // TODO: don't know where / how this is used. Looks like it is never used anywhere
-      payload: Bytes?,
-      arrivalTimestamp: Optional<UInt64>?,
-    ): PreparedGossipMessage = MaruPreparedGossipMessage(Bytes.fromHexString("deadbaaf"), Optional.empty())
-
-    override fun handleMessage(message: PreparedGossipMessage?): SafeFuture<ValidationResult> {
-      var data: Bytes?
-      message.let {
-        data = message!!.originalMessage
-      }
-      // at this point we have to validate the message (will only be further distributed if valid)
-      // at this point we should also (asynchonously) do what needs to be done with the data we received
-      dataFuture.complete(data)
-      return if (data!!.equals(Bytes.fromHexString(ORIGINAL_MESSAGE))) {
-        SafeFuture.completedFuture(ValidationResult.Valid)
-      } else {
-        SafeFuture.completedFuture(ValidationResult.Invalid)
-      }
-    }
-
-    override fun getMaxMessageSize(): Int = 43434343 // TODO: what is a good max size here? 10MB?
-  }
-
-  class MaruPreparedGossipMessage(
-    private val origMessage: Bytes,
-    private val arrTimestamp: Optional<UInt64>,
-  ) : PreparedGossipMessage {
-    override fun getMessageId(): Bytes = Bytes.of(42)
-
-    override fun getDecodedMessage(): PreparedGossipMessage.DecodedMessageResult =
-      PreparedGossipMessage.DecodedMessageResult.successful(origMessage)
-
-    override fun getOriginalMessage(): Bytes = origMessage
-
-    override fun getArrivalTimestamp(): Optional<UInt64> = arrTimestamp
-  }
-
-  companion object {
-    var rpcMethod = MaruRpcMethod()
-  }
-
-  class MaruRpcResponseHandler : RpcResponseHandler<Bytes> {
-    val future = SafeFuture<Bytes>()
-
-    override fun onResponse(response: Bytes?): SafeFuture<*> {
-      future.complete(response)
-      return SafeFuture.completedFuture(response)
-    }
-
-    override fun onCompleted(error: Optional<out Throwable>?): Unit = throw error!!.get()
-
-    fun response(): SafeFuture<Bytes> = future
   }
 }
